requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pytest
httpx

argon2-cffi
python-jose[cryptography]
email-validator
redis

numpy
scikit-learn
python-multipart
joblib

gunicorn